requirements.txt
# Core Flask
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# Database
psycopg2-binary
SQLAlchemy==2.0.36

# Authentication
PyJWT==2.9.0
bcrypt==4.2.0

# Deployment & Middlewares
gunicorn==22.0.0

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-flask==1.3.0
factory-boy==3.3.1
Faker==30.8.1

# Utilities
python-dateutil==2.9.0
Flask-CORS==4.0.0
